requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
docopt==0.6.2
frozenlist==1.3.0
idna==3.3
motor==3.0.0
multidict==6.0.2
pipreqs==0.4.11
psutil==5.9.1
pyaes==1.6.1
pyasn1==0.4.8
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
regex==2022.6.2
requests==2.28.0
reticker==3.1.3
rsa==4.8
speedtest-cli==2.1.3
urllib3==1.26.9
wincertstore==0.2
yarg==0.1.9
yarl==1.7.2
